added new parsing
