+++ src/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 11:46:46 by wnocchi           #+#    #+#             */
-/*   Updated: 2024/08/01 16:35:13 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/05 11:48:34 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
