+++ src/exec/child.c
@@ -6,7 +6,7 @@
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 10:33:48 by wnocchi           #+#    #+#             */
-/*   Updated: 2024/08/01 11:48:24 by wnocchi          ###   ########.fr       */
+/*   Updated: 2024/08/05 08:48:44 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ int	redirect_fd(t_msh *msh)
 	}
 	if (dup_in_fd(msh))
 		return (close_files(msh), 1);
-	if(msh->here_doc == 1)
+	if (msh->here_doc == 1)
 		unlink(msh->infile);
 	close_pipes(msh);
 	return (0);
