+++ src/parsing/fill_msh.c
@@ -1,17 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parsing.c                                          :+:      :+:    :+:   */
+/*   fill_msh.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/25 14:02:41 by tchalaou          #+#    #+#             */
-/*   Updated: 2024/07/31 14:31:17 by wnocchi          ###   ########.fr       */
+/*   Created: 2024/07/24 17:42:48 by tchalaou          #+#    #+#             */
+/*   Updated: 2024/08/05 09:02:22 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
-
 
 void	fill_command(t_msh *msh, t_token **token)
 {
@@ -29,54 +28,44 @@ void	fill_command(t_msh *msh, t_token **token)
 	msh->cmd[++i] = NULL;
 }
 
-void	fill_smaller(t_msh *msh, t_token **token)
+int	fill_smaller(t_msh *msh, t_token **token)
 {
 	*token = (*token)->next;
 	if (!*token)
-	{
-		printf("msh: parsing error\n");
-		return ;
-	}
+		return (printf("msh: parsing error\n"), 1);
 	if ((*token)->id == SMALLER)
 	{
 		*token = (*token)->next;
 		msh->here_doc = 1;
 	}
 	if (!*token)
-	{
-		printf("msh: parsing error\n");
-		return ;
-	}
+		return (printf("msh: parsing error\n"), 1);
 	if ((*token)->id == WORD)
 		msh->infile = ft_strdup((*token)->word);
 	else
-		printf("msh: parsing error\n");
+		return (printf("msh: parsing error\n"), 1);
 	*token = (*token)->next;
+	return (0);
 }
 
-void	fill_bigger(t_msh *msh, t_token **token)
+int	fill_bigger(t_msh *msh, t_token **token)
 {
 	*token = (*token)->next;
 	if (!*token)
-	{
-		printf("msh: parsing error\n");
-		return ;
-	}
+		return (printf("msh: parsing error\n"), 1);
 	if ((*token)->id == BIGGER)
 	{
 		*token = (*token)->next;
 		msh->append = 1;
 	}
 	if (!*token)
-	{
-		printf("msh: parsing error\n");
-		return ;
-	}
+		return (printf("msh: parsing error\n"), 1);
 	if ((*token)->id == WORD)
 		msh->outfile = ft_strdup((*token)->word);
 	else
-		printf("msh: parsing error\n");
+		return (printf("msh: parsing error\n"), 1);
 	*token = (*token)->next;
+	return (0);
 }
 
 void	fill_msh(t_msh *msh, t_token **token)
@@ -86,9 +75,15 @@ void	fill_msh(t_msh *msh, t_token **token)
 		if ((*token)->id == WORD)
 			fill_command(msh, token);
 		else if ((*token)->id == SMALLER)
-			fill_smaller(msh, token);
+		{
+			if (fill_smaller(msh, token))
+				break ;
+		}
 		else if ((*token)->id == BIGGER)
-			fill_bigger(msh, token);
+		{
+			if (fill_bigger(msh, token))
+				break ;
+		}
 		else if ((*token)->id == PIPE)
 		{
 			*token = (*token)->next;
@@ -98,66 +93,3 @@ void	fill_msh(t_msh *msh, t_token **token)
 			*token = (*token)->next;
 	}
 }
-
-t_msh	*parsing(t_token *token, t_env *env)
-{
-	t_msh	*msh;
-	t_msh	*add;
-	int		i;
-
-	msh = NULL;
-	i = -1;
-	while (token)
-	{
-		add = create_msh(++i, env);
-		if(!add)
-			return (NULL);
-		fill_msh(add, &token);
-		msh_add_back(&msh, add);
-	}
-	return (msh);
-}
-t_msh	*get_msh(char *line, t_env *env)
-{
-	t_token *token;
-	t_msh	*msh;
-	
-	token = lexing(line, env);
-	free(line);
-	if (!token)
-		return (NULL);
-	msh = parsing(token, env);
-	free_token(&token);
-	return (msh);
-}
-/*
-int	main(void)
-{
-	t_token	*token;
-	t_msh	*msh;
-	t_msh	*start;
-
-
-	token = lexing("< infile \"ls\" |'cat -e' Makefile > outfile | echo hello");
-	//token = lexing("cat -e | ls");
-	msh = parsing(token);
-	free_token(&token);
-	start = msh;
-	while (msh)
-	{
-		printf("index: %d\n", msh->index);
-		if (msh->cmd)
-		{
-			int	i = -1;
-			while (msh->cmd[++i])
-				printf("cmd[%d]: %s\n", i, msh->cmd[i]);
-		}
-		if (msh->infile)
-			printf("infile: %s\n", msh->infile);
-		if (msh->outfile)
-			printf("outfile: %s\n", msh->outfile);
-		msh = msh->next;
-	}
-	free_msh(&start);
-	return (0);
-}*/
