+++ src/parsing/fill_token.c
@@ -1,38 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexing.c                                           :+:      :+:    :+:   */
+/*   fill_token.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wnocchi <wnocchi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/25 13:51:23 by tchalaou          #+#    #+#             */
-/*   Updated: 2024/07/26 08:29:14 by wnocchi          ###   ########.fr       */
+/*   Created: 2024/07/24 17:36:44 by tchalaou          #+#    #+#             */
+/*   Updated: 2024/08/05 09:02:25 by wnocchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-char	*get_value_parsing(t_env *env, char *key)
-{
-	t_env	*current;
-	int	key_len;
-
-	current = env;
-	key_len = ft_strlen(key);
-	while (current)
-	{
-		if (!ft_strncmp(key, current->key, key_len) && !current->key[key_len])
-			return (current->value);
-		current = current->next;
-	}
-	return ("");
-}
-
 void	fill_value(t_token *token, char *line, int *i)
 {
 	char	*key;
 	char	*value;
-	char	*tmp;
 	int		k;
 
 	(*i)++;
@@ -47,17 +30,12 @@ void	fill_value(t_token *token, char *line, int *i)
 	}
 	key[k] = 0;
 	if (!ft_strlen(key))
-		value  = ft_strdup("$");
+		value = ft_strdup("$");
 	else
-		value = ft_strdup(get_value_parsing(token->env, key));
+		value = ft_strdup(get_env_value(token->env, key));
 	free(key);
 	if (value)
-	{
-		tmp = ft_strjoin(token->word, value);
-		free(value);
-		free(token->word);
-		token->word = tmp;
-	}
+		join_replace(&token->word, &value);
 }
 
 void	fill_word(t_token *token, char *line, int *i)
@@ -90,8 +68,7 @@ void	fill_quote(t_token *token, char *line, int *i)
 	int	len;
 
 	token->id = WORD;
-	(*i)++;
-	j = *i;
+	j = ++(*i);
 	len = 0;
 	while (line[j] && line[j++] != '\'')
 		len++;
@@ -100,13 +77,10 @@ void	fill_quote(t_token *token, char *line, int *i)
 		return ;
 	j = 0;
 	while (line[*i] && line[*i] != '\'')
-	{
-		token->word[j++] = line[*i];
-		(*i)++;
-	}
+		token->word[j++] = line[(*i)++];
 	token->word[j] = 0;
-	if (line[*i] == '\'')
-		(*i)++;
+	if (line[*i] != '\'')
+		(*i)--;
 }
 
 void	fill_doublequote(t_token *token, char *line, int *i)
@@ -115,8 +89,7 @@ void	fill_doublequote(t_token *token, char *line, int *i)
 	int	len;
 
 	token->id = WORD;
-	(*i)++;
-	j = *i;
+	j = ++(*i);
 	len = 0;
 	while (line[j] && line[j++] != '"')
 		len++;
@@ -129,16 +102,14 @@ void	fill_doublequote(t_token *token, char *line, int *i)
 		if (line[*i] == '$')
 		{
 			fill_value(token, line, i);
-			if (line[*i] == '"')
-				(*i)++;
-			return ;
+			break ;
 		}
-		token->word[j++] = line[*i];
-		(*i)++;
+		token->word[j++] = line[(*i)++];
 	}
-	token->word[j] = 0;
-	if (line[*i] == '"')
-		(*i)++;
+	if (j == len)
+		token->word[j] = 0;
+	if (line[*i] != '"')
+		(*i)--;
 }
 
 void	fill_token(t_token *token, char *line, int *i)
@@ -156,46 +127,7 @@ void	fill_token(t_token *token, char *line, int *i)
 	else if (line[*i] == '>')
 		token->id = BIGGER;
 	else if (line[*i] == '|')
-	{
-		if((size_t)*i != ft_strlen(line))
-			token->id = PIPE;
-	}
+		token->id = PIPE;
 	else
 		fill_word(token, line, i);
 }
-
-t_token	*lexing(char *line, t_env *env)
-{
-	t_token	*token;
-	t_token	*add;
-	int		i;
-
-	token = NULL;
-	i = -1;
-	while ((size_t)++i < ft_strlen(line))
-	{
-		add = create_token(env);
-		fill_token(add, line, &i);
-		token_add_back(&token, add);
-	}
-	return (token);
-}
-
-/*
-int	main()
-{
-	t_token	*token;
-	t_token *start;
-
-	token = lexing(" aa; <bb> | $cc ");
-	start = token;
-	while (token)
-	{
-		printf("id: %d\n", token->id);
-		if (token->word)
-			printf("word: %s\n", token->word);
-		token = token->next;
-	}
-	free_token(&start);
-	return (0);
-}*/
